feat: Add RestaurantVerticalList component for displaying a vertical list of restaurants
This commit adds a new component called RestaurantVerticalList, which is responsible for rendering a vertical list of restaurants. The component fetches the restaurant data from a remote server and displays it using a FlatList. Each restaurant item is rendered using the RestaurantItemVertical component.

The RestaurantVerticalList component is added to the main index file, allowing it to be displayed in the app. This change enhances the app by providing a way to view and interact with a vertical list of restaurants.